Move two handlers from user to mgmt context
Two handlers were incorrectly added to the user context instead
of the management context. These handlers are for mgmt plane resources:
globalRoleBinding and RoleTemplate, so they should be in the mgmt
plane.
