+++ pkg/controllers/management/auth/legacy_grb_cleaner.go
@@ -1,4 +1,4 @@
-package rbac
+package auth
 
 import (
 	"strings"
@@ -6,17 +6,22 @@ import (
 	"github.com/rancher/norman/lifecycle"
 	grbstore "github.com/rancher/rancher/pkg/api/store/globalrolebindings"
 	v3 "github.com/rancher/types/apis/management.cattle.io/v3"
+	"github.com/rancher/types/config"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
 type grbCleaner struct {
-	m *manager
+	clusterLister v3.ClusterLister
+	mgmt          *config.ManagementContext
 }
 
-func newLegacyGRBCleaner(m *manager) *grbCleaner {
-	return &grbCleaner{m: m}
+func newLegacyGRBCleaner(m *config.ManagementContext) *grbCleaner {
+	return &grbCleaner{
+		mgmt:          m,
+		clusterLister: m.Management.Clusters("").Controller().Lister(),
+	}
 }
 
 // this function addresses issues with grb not being cleaned up that was an issue from v2.2.3 - v2.2.8
@@ -38,7 +43,7 @@ func (p *grbCleaner) sync(key string, obj *v3.GlobalRoleBinding) (runtime.Object
 		obj.SetAnnotations(annotations)
 	}
 	obj.Annotations[grbstore.GrbVersion] = "true"
-	return p.m.workload.Management.Management.GlobalRoleBindings("").Update(obj)
+	return p.mgmt.Management.GlobalRoleBindings("").Update(obj)
 }
 
 func (p *grbCleaner) removeFinalizerFromNonExistentCluster(obj *v3.GlobalRoleBinding) (*v3.GlobalRoleBinding, error) {
@@ -55,7 +60,7 @@ func (p *grbCleaner) removeFinalizerFromNonExistentCluster(obj *v3.GlobalRoleBin
 		if strings.HasPrefix(f, lifecycle.ScopedFinalizerKey) {
 			s := strings.Split(f, "_")
 			// if cluster was not reported, its a deleted cluster and will cause the finalizer to hang in the future
-			if _, err = p.m.clusterLister.Get("", s[1]); errors.IsNotFound(err) {
+			if _, err = p.clusterLister.Get("", s[1]); errors.IsNotFound(err) {
 				finalizers = append(finalizers[:i], finalizers[i+1:]...)
 			}
 		}
