+++ pkg/controllers/management/auth/register.go
@@ -16,6 +16,8 @@ func RegisterEarly(ctx context.Context, management *config.ManagementContext, cl
 	n := newTokenController(management)
 	ua := newUserAttributeController(management)
 	s := newAuthSettingController(management)
+	grbLegacy := newLegacyGRBCleaner(management)
+	rtLegacy := newLegacyRTCleaner(management)
 
 	management.Management.ProjectRoleTemplateBindings("").AddLifecycle(ctx, ptrbMGMTController, prtb)
 	management.Management.ClusterRoleTemplateBindings("").AddLifecycle(ctx, ctrbMGMTController, crtb)
@@ -27,6 +29,8 @@ func RegisterEarly(ctx context.Context, management *config.ManagementContext, cl
 	management.Management.Tokens("").AddHandler(ctx, tokenController, n.sync)
 	management.Management.UserAttributes("").AddHandler(ctx, userAttributeController, ua.sync)
 	management.Management.Settings("").AddHandler(ctx, authSettingController, s.sync)
+	management.Management.GlobalRoleBindings("").AddHandler(ctx, "legacy-grb-cleaner", grbLegacy.sync)
+	management.Management.RoleTemplates("").AddHandler(ctx, "legacy-rt-cleaner", rtLegacy.sync)
 }
 
 func RegisterLate(ctx context.Context, management *config.ManagementContext) {
